Replace app/app.go with two files
