+++ node/bootstrap/bootstrap.go
@@ -1,32 +1,18 @@
-// Copyright (C) 2019-2025, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package app
+package bootstrap
 
 import (
 	"fmt"
-	"os"
-	"os/signal"
 	"sync"
-	"syscall"
 
 	"go.uber.org/zap"
 
 	"github.com/ava-labs/avalanchego/node"
-	"github.com/ava-labs/avalanchego/utils"
 	"github.com/ava-labs/avalanchego/utils/logging"
 	"github.com/ava-labs/avalanchego/utils/perms"
 	"github.com/ava-labs/avalanchego/utils/ulimit"
 
 	nodeconfig "github.com/ava-labs/avalanchego/config/node"
 )
-
-const Header = `     _____               .__                       .__
-    /  _  \___  _______  |  | _____    ____   ____ |  |__   ____    ,_ o
-   /  /_\  \  \/ /\__  \ |  | \__  \  /    \_/ ___\|  |  \_/ __ \   / //\,
-  /    |    \   /  / __ \|  |__/ __ \|   |  \  \___|   Y  \  ___/    \>> |
-  \____|__  /\_/  (____  /____(____  /___|  /\___  >___|  /\___  >    \\
-          \/           \/          \/     \/     \/     \/     \/`
 
 var _ App = (*app)(nil)
 
@@ -43,6 +29,14 @@ type App interface {
 	// ExitCode should only be called after [Start] returns. It
 	// should block until the application finishes
 	ExitCode() int
+}
+
+// app is a wrapper around a node that runs in this process
+type app struct {
+	node       *node.Node
+	log        logging.Logger
+	logFactory logging.Factory
+	exitWG     sync.WaitGroup
 }
 
 func New(config nodeconfig.Config) (App, error) {
@@ -84,56 +78,6 @@ func New(config nodeconfig.Config) (App, error) {
 		log:        log,
 		logFactory: logFactory,
 	}, nil
-}
-
-func Run(app App) int {
-	// start running the application
-	app.Start()
-
-	// register terminationSignals to kill the application
-	terminationSignals := make(chan os.Signal, 1)
-	signal.Notify(terminationSignals, syscall.SIGINT, syscall.SIGTERM)
-
-	stackTraceSignal := make(chan os.Signal, 1)
-	signal.Notify(stackTraceSignal, syscall.SIGABRT)
-
-	// start up a new go routine to handle attempts to kill the application
-	go func() {
-		for range terminationSignals {
-			app.Stop()
-			return
-		}
-	}()
-
-	// start a goroutine to listen on SIGABRT signals,
-	// to print the stack trace to standard error.
-	go func() {
-		for range stackTraceSignal {
-			fmt.Fprint(os.Stderr, utils.GetStacktrace(true))
-		}
-	}()
-
-	// wait for the app to exit and get the exit code response
-	exitCode := app.ExitCode()
-
-	// shut down the termination signal go routine
-	signal.Stop(terminationSignals)
-	close(terminationSignals)
-
-	// shut down the stack trace go routine
-	signal.Stop(stackTraceSignal)
-	close(stackTraceSignal)
-
-	// return the exit code that the application reported
-	return exitCode
-}
-
-// app is a wrapper around a node that runs in this process
-type app struct {
-	node       *node.Node
-	log        logging.Logger
-	logFactory logging.Factory
-	exitWG     sync.WaitGroup
 }
 
 // Start the business logic of the node (as opposed to config reading, etc).
